fix: move correctness tests from alloc_test.go to rule_test.go
CLAUDE.md requires unit tests in rule_test.go. TestCheck_BothViolations,
TestCheck_MultipleBareURLsPreSized, TestCheck_AllowMarkerDirectiveInMessage,
and TestCheck_CustomAllowMarker are correctness checks, not allocation gates,
and belong alongside the other rule unit tests.
